9_ML_System_Monitoring/monitoring_proj_evidently/fastapi/requirements.txt
fastapi
uvicorn
pillow
prometheus_client
python-multipart
psycopg2-binary
scikit-learn
pandas
evidently==0.6.6